// mobile/gen/models/CreateAppointment.ts
export const createAppointment201StatusEnum = {
  AGENDADA: "AGENDADA",
  CANCELADA: "CANCELADA",
  REAGENDADA: "REAGENDADA",
  CONCLUIDA: "CONCLUIDA",
} as const;

export type CreateAppointment201StatusEnumKey =
  (typeof createAppointment201StatusEnum)[keyof typeof createAppointment201StatusEnum];

export const createAppointment201PriorityEnum = {
  BAIXA: "BAIXA",
  MEDIA: "MEDIA",
  ALTA: "ALTA",
} as const;

export type CreateAppointment201PriorityEnumKey =
  (typeof createAppointment201PriorityEnum)[keyof typeof createAppointment201PriorityEnum];

export const createAppointment201SourceEnum = {
  TRIAGEM: "TRIAGEM",
  DIRECTA: "DIRECTA",
} as const;

export type CreateAppointment201SourceEnumKey =
  (typeof createAppointment201SourceEnum)[keyof typeof createAppointment201SourceEnum];

/**
 * @description appointment created successfully
 */
export type CreateAppointment201 = {
  /**
   * @type string
   */
  id: string;
  /**
   * @type string, date-time
   */
  date: string;
  /**
   * @type string
   */
  patientId: string;
  /**
   * @type string
   */
  specialtyId: string;
  /**
   * @type string
   */
  consultationTypeId: string;
  /**
   * @type string | undefined
   */
  doctorId?: string;
  /**
   * @type string | undefined
   */
  triageId?: string;
  /**
   * @type string | undefined
   */
  notes?: string;
  /**
   * @default "AGENDADA"
   * @type string
   */
  status: CreateAppointment201StatusEnumKey;
  /**
   * @default "MEDIA"
   * @type string
   */
  priority: CreateAppointment201PriorityEnumKey;
  /**
   * @default "DIRECTA"
   * @type string
   */
  source: CreateAppointment201SourceEnumKey;
  /**
   * @type string, date-time
   */
  createdAt: string;
  /**
   * @type string, date-time
   */
  updatedAt: string;
};

/**
 * @description Invalid input data provided
 */
export type CreateAppointment400 = any;

/**
 * @description Authentication required
 */
export type CreateAppointment401 = any;

/**
 * @description Insufficient permissions
 */
export type CreateAppointment403 = any;

export const createAppointmentMutationRequestStatusEnum = {
  AGENDADA: "AGENDADA",
  CANCELADA: "CANCELADA",
  REAGENDADA: "REAGENDADA",
  CONCLUIDA: "CONCLUIDA",
} as const;

export type CreateAppointmentMutationRequestStatusEnumKey =
  (typeof createAppointmentMutationRequestStatusEnum)[keyof typeof createAppointmentMutationRequestStatusEnum];

export const createAppointmentMutationRequestPriorityEnum = {
  BAIXA: "BAIXA",
  MEDIA: "MEDIA",
  ALTA: "ALTA",
} as const;

export type CreateAppointmentMutationRequestPriorityEnumKey =
  (typeof createAppointmentMutationRequestPriorityEnum)[keyof typeof createAppointmentMutationRequestPriorityEnum];

export const createAppointmentMutationRequestSourceEnum = {
  TRIAGEM: "TRIAGEM",
  DIRECTA: "DIRECTA",
} as const;

export type CreateAppointmentMutationRequestSourceEnumKey =
  (typeof createAppointmentMutationRequestSourceEnum)[keyof typeof createAppointmentMutationRequestSourceEnum];

export type CreateAppointmentMutationRequest = {
  date: string | string;
  /**
   * @type object | undefined
   */
  patient?: {
    /**
     * @minLength 1
     * @type string
     */
    id: string;
  };
  /**
   * @type object | undefined
   */
  specialty?: {
    /**
     * @minLength 1
     * @type string
     */
    id: string;
  };
  /**
   * @type object | undefined
   */
  consultationType?: {
    /**
     * @minLength 1
     * @type string
     */
    id: string;
  };
  /**
   * @type object | undefined
   */
  doctor?: {
    /**
     * @minLength 1
     * @type string
     */
    id: string;
  };
  /**
   * @type object | undefined
   */
  triage?: {
    /**
     * @minLength 1
     * @type string
     */
    id: string;
  };
  /**
   * @minLength 1
   * @type string | undefined
   */
  notes?: string;
  /**
   * @type string | undefined
   */
  status?: CreateAppointmentMutationRequestStatusEnumKey;
  /**
   * @type string | undefined
   */
  priority?: CreateAppointmentMutationRequestPriorityEnumKey;
  /**
   * @type string | undefined
   */
  source?: CreateAppointmentMutationRequestSourceEnumKey;
};