// mobile/gen/models/FindAppointments.ts
export const dataStatusEnum = {
  AGENDADA: "AGENDADA",
  CANCELADA: "CANCELADA",
  REAGENDADA: "REAGENDADA",
  CONCLUIDA: "CONCLUIDA",
} as const;

export type DataStatusEnumKey =
  (typeof dataStatusEnum)[keyof typeof dataStatusEnum];

export const dataPriorityEnum = {
  BAIXA: "BAIXA",
  MEDIA: "MEDIA",
  ALTA: "ALTA",
} as const;

export type DataPriorityEnumKey =
  (typeof dataPriorityEnum)[keyof typeof dataPriorityEnum];

export const dataSourceEnum = {
  TRIAGEM: "TRIAGEM",
  DIRECTA: "DIRECTA",
} as const;

export type DataSourceEnumKey =
  (typeof dataSourceEnum)[keyof typeof dataSourceEnum];

/**
 * @description List of appointments retrieved successfully
 */
export type FindAppointments200 = {
  /**
   * @description Total number of records matching the criteria
   * @type integer | undefined
   */
  total?: number;
  /**
   * @description Number of records returned in current page
   * @type integer | undefined
   */
  results?: number;
  /**
   * @type array | undefined
   */
  data?: {
    /**
     * @type string
     */
    id: string;
    /**
     * @type string, date-time
     */
    date: string;
    /**
     * @type string
     */
    patientId: string;
    /**
     * @type string
     */
    specialtyId: string;
    /**
     * @type string
     */
    consultationTypeId: string;
    /**
     * @type string | undefined
     */
    doctorId?: string;
    /**
     * @type string | undefined
     */
    triageId?: string;
    /**
     * @type string | undefined
     */
    notes?: string;
    /**
     * @default "AGENDADA"
     * @type string
     */
    status: DataStatusEnumKey;
    /**
     * @default "MEDIA"
     * @type string
     */
    priority: DataPriorityEnumKey;
    /**
     * @default "DIRECTA"
     * @type string
     */
    source: DataSourceEnumKey;
    /**
     * @type string, date-time
     */
    createdAt: string;
    /**
     * @type string, date-time
     */
    updatedAt: string;
  }[];
};